refactor duplicated function
